OAK-5902 - Cold standby should allow syncing of blobs bigger than 2.2 GB Implemented chunking algorithm on client and server
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1804511 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-segment-tar/src/main/java/org/apache/jackrabbit/oak/segment/standby/store/CommunicationObserver.java
@@ -101,7 +101,7 @@ public class CommunicationObserver {
         partnerDetails.put(client, m);
     }
 
-    public void didSendBinariesBytes(String client, int size) {
+    public void didSendBinariesBytes(String client, long size) {
         log.debug("Binary with size {} sent to client {}", size, client);
         CommunicationPartnerMBean m = partnerDetails.get(client);
         m.onBinarySent(size);
